fix: gets everything running smoothly
