+++ apps/ai-hero/src/lib/posts.ts
@@ -68,3 +68,19 @@ export const PostUpdateSchema = z.object({
 })
 
 export type PostUpdate = z.infer<typeof PostUpdateSchema>
+
+export const NewPostInputSchema = z.object({
+	title: z.string().min(1, 'Title is required'),
+	videoResourceId: z.string().optional(),
+	postType: z.enum([
+		'lesson',
+		'podcast',
+		'tip',
+		'course',
+		'playlist',
+		'article',
+	]),
+	createdById: z.string(),
+})
+
+export type NewPostInput = z.infer<typeof NewPostInputSchema>
